Convert from float to double (to avoid rounding breaking unit tests!)

+++ src/jpstrack/model/Reading.java
@@ -4,18 +4,18 @@ import java.util.Date;
 
 public class Reading {
 	long time;
-	float latitude;
-	float longitude;
-	public float getLatitude() {
+	double latitude;
+	double longitude;
+	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
-	public float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 	public long getTime() {
